snapd-client, tests: Don't iterate a main context we didn't create
Following

  https://developer.gnome.org/programming-guidelines/unstable/main-contexts.html.en#using-gmaincontext-in-a-library

We shouldn't be iterating a GMainContext that might affect the
application which is using us. Try to fix that by creating our own and
only using that one. Also try to not expose this to the application, by
not setting it as the thread default. We're not in charge of the
lifecycle of our thread.

The testsuite was actually *relying* on the library iterating the
thread-default GMainContext. I think that was bad behaviour, so I
changed it to not rely on that - it creates a worker thread for the mock
snapd and runs that thread's GMainContext.