Add nanosecond accuracy support
To work as expected, the NOTICES snapd API requires nanosecond
accuracy. Unfortunately, at the moment GDateTime has only
microsecond accuracy, which means that some kind of trick is
required to be able to pass nanosecond accuracy.

This patch implements that trick, by adding a method that allows
to get the seconds value of the last occurred date, but with
nanosecond accuracy, and use that value in the calls for reading
more notices, in the AFTER field.