Object for monitoring new notices
This object allows to easily monitor new notices. It creates a
new client connection (because the call is "blocking", in the
sense that, when a client is asking for new notices in a socket,
that socket can't accept new commands until a new notice is
returned or the timeout has expired), and asks for new notices
with a very big timeout. Whenever a new notice arrives, it sends
a signal with that notice, and waits again for new notices.

This PR requires #150, so, until that is merged, must be marked
as draft.