Add snapd_notice_get_XXXX2 versions for getters
Several getters were returning a reference of the data instead
of just a pointer; this is, they incremented the refcount of the
returned element. This isn't what the rest of the API does, so
for uniformity, this PR adds new versions of the getters and
deprecates the old ones.

Fix #170