Fix GIR comments (#191)
The autoformatting has the problem of cutting comments in several lines, which is a problem for the GIR parser if a GIR line with parentheses is cut and the opening and closing parentheses are in different lines.

This change instructs clang-format to not reflow comments. It also fixes the GIR errors.