Added test support for "after" (#197)
* Added test support for "after"

The mock-snapd class allows to simulate receiving notices, but
until now it didn't honor the "after" field, so it always
returned all the notices.

This patch changes this, also allowing to test that the "after"
parameter is correctly parsed. This also allowed to find a bug
when requesting notices after a specific date when using
GLib 2.65 or previous.