Fix warning
The snapd_notice_get_notice_type() method returns a notice type,
which is an `enum`. But the return type is defined as `const`,
which causes a warning in the compiler when using the option
`-Wignored-qualifiers`.

This patch removes the unneeded `const` qualifier to remove this
warning.

Fix https://github.com/canonical/snapd-glib/issues/200