Remove support for Ubuntu 18.04
The checkout@v3 action relies on node20, which requires a
version of GLibc newer than the one distributed in Ubuntu
18.04. Until now we used a workaround to be able to still run
the action and, thus, execute the tests, but that workaround
ceased to work and now neither the current code, nor any new
PR do pass the continuous integration tests in github.

Since Ubuntu 18.04 reached end-of-life in May 2023, it makes
sense to remove support.