Use the new abstract socket inside a snap
The snap team has changed the socket path used to communicate
with snapd from inside a snap to an abstract socket.

Although snapd-glib didn't manage the case of being used inside
a snap, this change gives us the opportunity to both fix this,
and ensure compatibility with the new socket and old systems.

This patch checks if the library is being run from inside a
snap. If that's not the case, the old, classic socket
/run/snapd.socket will be used. But if that's the case, the new
abstract socket @/snapd/snapd-snap.socket will be tested first,
and only if it's not possible to connect, will the old "inside
snap" socket, /run/snapd-snap.socket, be used.